initial
InputBinding rewrite to allow mouse and keyboard inputs.  InputTrigger handles the assignments. simplified many calls to InputBinding.Process for methods that are now unnecessary and can be handled entirely from the control or window context. Added preprocessor KeyboardKeysString and MouseButtonString to enable constant use for defaults in InputOptions.  Changed a few things in EditorViewport to show proof of concept.