much progress, see description
id like to get everything fixed before push, but i need to get this in there.
this moves alot of the widgets into a module, has the beginnings of removing the Input struct that should be elsewhere (I make the argument it should be a separate namespace). Fixes to InputBindingList  to ensure that the picked binding is the one that matches most completely instead of the first one that satisfies the key combination. still has many broken features, as well as cacheing of settings. features were removed for expediency from EditorViewport including gamepad and trackpad code. can be added back in at a later date. in order to unify the widget usage widgets will need to be removed from other classes extending EditorViewport