refactored InputBindingList
greatly simplified this with a new struct that decouples the input binding from the actions it triggers. this allows setting a ground truth for input bindings. made an enum as well as a static dictionary for the inputoptions to easily reference the bindings, allowing propagation to any part of the codebase that uses the bindings.