Improvements to the fbx importer
- FBX files with an Up axis different from the Y axis are now rotated and
positioned correctly in the engine.
- Geometric transformation has been incorporated for meshes, and the copying
of this data has been corrected.
- Fixed a bug where the importer did not correctly detect the axes of the
FBX file.
    “UpAxi used the UpVector enum (from ofbx), but for some reason it
works better to use the integer values directly.
    “GlobalSettings” is now created from “OpenFbxImporter” and called from
there.
- “ConvertRootRotation” has been replaced by a "ToFlax" matrix
used to correctly rotate and position meshes, not at the root level.
- The main functions of “OpenFbxImporter” have been moved outside. This
struct has been left as an auxiliary data type.
- Two new structs, “AxisSystem” and “CoordSystem,” have been created,
which are created in “OpenFbxImporter.”
The parsing of the coordinate system has been moved to its own function.
-A function inside "CoordSystem" has been created to evaluate the handedness
of the FBX file.
- The Scene object destructor is now called.